ASTUDIO_Project/requirements.txt
Flask==2.2.2
openai==<version>  # Replace <version> with the correct version (e.g., 0.27.0)
python-dotenv==0.21.0
spacy==3.4.0
pdfplumber==0.5.28
pytesseract==0.3.9
python-docx==0.8.11
PyMuPDF==1.18.19
numpy==1.24.2
pdfminer.six==20201018
